requirements.txt
flask==3.1.1
python-dotenv==1.1.0
requests==2.32.2
pymysql==1.1.1
cryptography==45.0.4